X015 owner view constraints

+++ clubs/views/member_management_views.py
@@ -15,13 +15,19 @@ from clubs.models import Membership, Club, Ban
 @login_required
 @membership_exists
 def kick_member(request, member_id):
-    """Kick a given member from their club."""
+    """Allow an owner or officer to kick a given member from their club."""
     current_user = request.user
     member = Membership.objects.get(id=member_id)
     club = member.club
     if is_user_owner_of_club(current_user, club) or is_user_officer_of_club(current_user, club):
-        Membership.objects.filter(id=member_id).delete()
-        messages.warning(request, '@' + member.user.username + ' was kicked from the club.')
+        if not is_user_officer_of_club(member.user, club):
+            if not is_user_owner_of_club(member.user, club):
+                Membership.objects.filter(id=member_id).delete()
+                messages.warning(request, '@' + member.user.username + ' was kicked from the club.')
+            else:
+                messages.error(request, 'You are the owner. You cannot kick yourself from your club. Transfer ownership before leaving or delete the club.')
+        else:
+            messages.error(request, 'You cannot kick a officer, demote them first (owner only).')
     else:
         messages.error(request, 'Only the owners and officers can kick members from a club.')
     return redirect('members_list', club_id=club.id)
@@ -29,15 +35,18 @@ def kick_member(request, member_id):
 @login_required
 @membership_exists
 def ban_member(request, member_id):
-    """Ban a given member from their club."""
+    """Allow the owner to ban a given member from their club."""
     current_user = request.user
     member = Membership.objects.get(id=member_id)
     club = member.club
     if is_user_owner_of_club(current_user, club):
         if not is_user_officer_of_club(member.user, club): #Owners can only ban members, not officers.
-            Ban.objects.create(club=club, user=member.user)
-            Membership.objects.filter(id=member_id).delete()
-            messages.warning(request, '@' + member.user.username + ' was banned from the club.')
+            if not is_user_owner_of_club(member.user, club): #An owner cannot ban themselves.
+                Ban.objects.create(club=club, user=member.user)
+                Membership.objects.filter(id=member_id).delete()
+                messages.warning(request, '@' + member.user.username + ' was banned from the club.')
+            else:
+                messages.error(request, "You are the owner. You cannot ban yourself from your club.")
         else: #Tried to ban an officer.
             messages.error(request, 'You cannot ban an officer. Demote them first.')
     else: #Not the owner
@@ -60,7 +69,7 @@ def banned_members(request, club_id):
 @login_required
 @ban_exists
 def unban_member(request, ban_id):
-    """Revoke a given ban from their club."""
+    """Allow the owner to revoke a given ban from their club."""
     current_user = request.user
     ban = Ban.objects.get(id=ban_id)
     club = ban.club
@@ -69,8 +78,8 @@ def unban_member(request, ban_id):
         Ban.objects.filter(id=ban_id).delete()
         messages.warning(request, '@' + ban.user.username + ' was unban from the club.')
     else:
-        member = Membership.objects.get(club=club, is_owner=True)
         messages.error(request, 'Only the owner can unban members.')
+
     return redirect('banned_members', club_id=club.id)
 
 @login_required
@@ -80,11 +89,18 @@ def promote_member_to_officer(request, member_id):
     member = Membership.objects.get(id = member_id)
     club = member.club
     if is_user_owner_of_club(request.user, club):
-        member.is_officer = True
-        member.save() # Or database won't update.
-        messages.success(request, '@' + member.user.username + ' was promoted to officer.')
-    else: # Access denied, member isn't owner
+        if not is_user_owner_of_club(member.user, club):
+            if not is_user_officer_of_club(member.user, club):
+                member.is_officer = True
+                member.save() # Or database won't update.
+                messages.success(request, '@' + member.user.username + ' was promoted to officer.')
+            else: #Access denied owner is trying to promote an officer.
+                messages.warning(request, '@' + member.user.username + ' is already an officer.')
+        else: #Access denied owner is trying to promote themselves.
+            messages.error(request, 'You are the owner. You cannot promote yourself to officer.')
+    else: # Access denied, member isn't owner.
         messages.error(request, 'Only the owner can promote members.')
+
     return redirect('members_list', club_id=club.id)
 
 @login_required
@@ -94,9 +110,16 @@ def demote_officer_to_member(request, member_id):
     member = Membership.objects.get(id = member_id)
     club = member.club
     if is_user_owner_of_club(request.user, club) :
-        member.is_officer = False
-        member.save() # Or database won't update.
-        messages.warning(request, '@' + member.user.username + ' was demoted.')
+        if not is_user_owner_of_club(member.user, club):
+            if is_user_officer_of_club(member.user, club):
+                member.is_officer = False
+                member.save() # Or database won't update.
+                messages.warning(request, '@' + member.user.username + ' was demoted.')
+            else: #Access denied, trying to demote non-officer.
+                messages.warning(request, '@' + member.user.username + ' is not an officer.')
+        else: #Access denied, owner is trying to demote themselves.
+            messages.error(request, 'You are the owner. You cannot demote yourself. You can transfer ownership from the member list.')
     else: # Access denied, member isn't owner
         messages.error(request, 'Only the owner can demote members.')
+
     return redirect('members_list', club_id=club.id)
